--- alembic/versions/164db40f2a51_added_is_superadmin_to_user_table.py
@@ -1,30 +1,0 @@
-"""added is_superadmin to user table
-
-Revision ID: 164db40f2a51
-Revises: 3e6101c469ca
-Create Date: 2024-07-23 11:17:35.350313
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '164db40f2a51'
-down_revision: Union[str, None] = '3e6101c469ca'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('users', sa.Column('is_superadmin', sa.Boolean(), server_default=sa.text('false'), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('users', 'is_superadmin')
-    # ### end Alembic commands ###
