--- alembic/versions/69b3a024890e_create_tables.py
@@ -1,222 +1,0 @@
-"""create tables
-
-Revision ID: 69b3a024890e
-Revises: 
-Create Date: 2024-07-22 22:26:26.382049
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision: str = '69b3a024890e'
-down_revision: Union[str, None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('newsletters',
-    sa.Column('email', sa.String(length=150), nullable=False),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('email')
-    )
-    op.create_index(op.f('ix_newsletters_id'), 'newsletters', ['id'], unique=False)
-    op.create_table('organizations',
-    sa.Column('name', sa.String(length=50), nullable=False),
-    sa.Column('description', sa.Text(), nullable=True),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_organizations_id'), 'organizations', ['id'], unique=False)
-    op.create_table('permissions',
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_permissions_id'), 'permissions', ['id'], unique=False)
-    op.create_index(op.f('ix_permissions_name'), 'permissions', ['name'], unique=False)
-    op.create_table('products',
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('description', sa.Text(), nullable=True),
-    sa.Column('price', sa.Numeric(), nullable=False),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_products_id'), 'products', ['id'], unique=False)
-    op.create_table('users',
-    sa.Column('username', sa.String(length=50), nullable=False),
-    sa.Column('email', sa.String(length=100), nullable=False),
-    sa.Column('password', sa.String(length=255), nullable=False),
-    sa.Column('first_name', sa.String(length=50), nullable=True),
-    sa.Column('last_name', sa.String(length=50), nullable=True),
-    sa.Column('is_active', sa.Boolean(), server_default=sa.text('true'), nullable=True),
-    sa.Column('is_admin', sa.Boolean(), server_default=sa.text('false'), nullable=True),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('email'),
-    sa.UniqueConstraint('username')
-    )
-    op.create_index(op.f('ix_users_id'), 'users', ['id'], unique=False)
-    op.create_table('waitlist_users',
-    sa.Column('email', sa.String(length=100), nullable=False),
-    sa.Column('full_name', sa.String(length=100), nullable=False),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('email')
-    )
-    op.create_index(op.f('ix_waitlist_users_id'), 'waitlist_users', ['id'], unique=False)
-    op.create_table('blogs',
-    sa.Column('author_id', sa.String(), nullable=False),
-    sa.Column('title', sa.String(length=100), nullable=False),
-    sa.Column('content', sa.Text(), nullable=True),
-    sa.Column('image_url', sa.String(length=100), nullable=True),
-    sa.Column('tags', postgresql.ARRAY(sa.String(length=20)), nullable=True),
-    sa.Column('is_deleted', sa.Boolean(), nullable=False),
-    sa.Column('excerpt', sa.String(length=500), nullable=True),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['author_id'], ['users.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_blogs_id'), 'blogs', ['id'], unique=False)
-    op.create_table('invitations',
-    sa.Column('user_id', sa.String(), nullable=False),
-    sa.Column('organization_id', sa.String(), nullable=False),
-    sa.Column('expires_at', sa.DateTime(timezone=True), nullable=False),
-    sa.Column('is_valid', sa.Boolean(), nullable=True),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['organization_id'], ['organizations.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_invitations_id'), 'invitations', ['id'], unique=False)
-    op.create_table('jobs',
-    sa.Column('user_id', sa.String(), nullable=False),
-    sa.Column('title', sa.String(length=255), nullable=False),
-    sa.Column('description', sa.Text(), nullable=False),
-    sa.Column('location', sa.String(length=255), nullable=True),
-    sa.Column('salary', sa.Numeric(precision=10, scale=2), nullable=True),
-    sa.Column('job_type', sa.String(length=50), nullable=True),
-    sa.Column('company_name', sa.String(length=255), nullable=True),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_jobs_id'), 'jobs', ['id'], unique=False)
-    op.create_table('profiles',
-    sa.Column('user_id', sa.String(), nullable=False),
-    sa.Column('bio', sa.Text(), nullable=True),
-    sa.Column('phone_number', sa.String(length=50), nullable=True),
-    sa.Column('avatar_url', sa.String(length=100), nullable=True),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('user_id')
-    )
-    op.create_index(op.f('ix_profiles_id'), 'profiles', ['id'], unique=False)
-    op.create_table('roles',
-    sa.Column('role_name', sa.String(), nullable=False),
-    sa.Column('organization_id', sa.String(), nullable=False),
-    sa.Column('is_active', sa.Boolean(), nullable=True),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['organization_id'], ['organizations.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_roles_id'), 'roles', ['id'], unique=False)
-    op.create_index(op.f('ix_roles_role_name'), 'roles', ['role_name'], unique=False)
-    op.create_table('subscriptions',
-    sa.Column('user_id', sa.String(), nullable=False),
-    sa.Column('plan', sa.String(length=50), nullable=False),
-    sa.Column('is_active', sa.Boolean(), nullable=True),
-    sa.Column('start_date', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('end_date', sa.DateTime(timezone=True), nullable=True),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_subscriptions_id'), 'subscriptions', ['id'], unique=False)
-    op.create_table('user_organization',
-    sa.Column('user_id', sa.String(), nullable=False),
-    sa.Column('organization_id', sa.String(), nullable=False),
-    sa.ForeignKeyConstraint(['organization_id'], ['organizations.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('user_id', 'organization_id')
-    )
-    op.create_table('role_permission',
-    sa.Column('role_id', sa.String(), nullable=False),
-    sa.Column('permission_id', sa.String(), nullable=False),
-    sa.ForeignKeyConstraint(['permission_id'], ['permissions.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['role_id'], ['roles.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('role_id', 'permission_id')
-    )
-    op.create_table('user_role',
-    sa.Column('user_id', sa.String(), nullable=False),
-    sa.Column('role_id', sa.String(), nullable=False),
-    sa.ForeignKeyConstraint(['role_id'], ['roles.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('user_id', 'role_id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('user_role')
-    op.drop_table('role_permission')
-    op.drop_table('user_organization')
-    op.drop_index(op.f('ix_subscriptions_id'), table_name='subscriptions')
-    op.drop_table('subscriptions')
-    op.drop_index(op.f('ix_roles_role_name'), table_name='roles')
-    op.drop_index(op.f('ix_roles_id'), table_name='roles')
-    op.drop_table('roles')
-    op.drop_index(op.f('ix_profiles_id'), table_name='profiles')
-    op.drop_table('profiles')
-    op.drop_index(op.f('ix_jobs_id'), table_name='jobs')
-    op.drop_table('jobs')
-    op.drop_index(op.f('ix_invitations_id'), table_name='invitations')
-    op.drop_table('invitations')
-    op.drop_index(op.f('ix_blogs_id'), table_name='blogs')
-    op.drop_table('blogs')
-    op.drop_index(op.f('ix_waitlist_users_id'), table_name='waitlist_users')
-    op.drop_table('waitlist_users')
-    op.drop_index(op.f('ix_users_id'), table_name='users')
-    op.drop_table('users')
-    op.drop_index(op.f('ix_products_id'), table_name='products')
-    op.drop_table('products')
-    op.drop_index(op.f('ix_permissions_name'), table_name='permissions')
-    op.drop_index(op.f('ix_permissions_id'), table_name='permissions')
-    op.drop_table('permissions')
-    op.drop_index(op.f('ix_organizations_id'), table_name='organizations')
-    op.drop_table('organizations')
-    op.drop_index(op.f('ix_newsletters_id'), table_name='newsletters')
-    op.drop_table('newsletters')
-    # ### end Alembic commands ###
