+++ alembic/versions/ced363af8b02_added_org_role_tables.py
@@ -1,8 +1,8 @@
-"""added is_deleetd to user table
+"""added org_role tables
 
-Revision ID: 3e6101c469ca
-Revises: 69b3a024890e
-Create Date: 2024-07-23 09:14:07.716941
+Revision ID: ced363af8b02
+Revises: 3d7695723bfa
+Create Date: 2024-07-24 06:55:50.835200
 
 """
 from typing import Sequence, Union
@@ -12,19 +12,19 @@ import sqlalchemy as sa
 
 
 # revision identifiers, used by Alembic.
-revision: str = '3e6101c469ca'
-down_revision: Union[str, None] = '69b3a024890e'
+revision: str = 'ced363af8b02'
+down_revision: Union[str, None] = '3d7695723bfa'
 branch_labels: Union[str, Sequence[str], None] = None
 depends_on: Union[str, Sequence[str], None] = None
 
 
 def upgrade() -> None:
     # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('users', sa.Column('is_deleted', sa.Boolean(), server_default=sa.text('false'), nullable=True))
+    pass
     # ### end Alembic commands ###
 
 
 def downgrade() -> None:
     # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('users', 'is_deleted')
+    pass
     # ### end Alembic commands ###
