+++ api/v1/models/__init__.py
@@ -1,11 +1,11 @@
-from api.v1.models.user import User, WaitlistUser
+from api.v1.models.user import User
 from api.v1.models.org import Organization
-from api.v1.models.profile import Profile
-from api.v1.models.product import Product
-from api.v1.models.subscription import Subscription
-from api.v1.models.blog import Blog
-from api.v1.models.job import Job
-from api.v1.models.invitation import Invitation
-from api.v1.models.role import Role
-from api.v1.models.permission import Permission
-from api.v1.models.newsletter import Newsletter
+# from api.v1.models.profile import Profile
+# from api.v1.models.product import Product
+# from api.v1.models.subscription import Subscription
+# from api.v1.models.blog import Blog
+# from api.v1.models.job import Job
+# from api.v1.models.invitation import Invitation
+# from api.v1.models.role import Role
+# from api.v1.models.permission import Permission
+# from api.v1.models.newsletter import Newsletter
