+++ api/v1/models/newsletter.py
@@ -1,15 +1,15 @@
-from sqlalchemy import Column, String
-from uuid import uuid4
-from datetime import datetime
-from api.db.database import Base
-from api.v1.models.base_model import BaseTableModel
+# from sqlalchemy import Column, String
+# from uuid import uuid4
+# from datetime import datetime
+# from api.db.database import Base
+# from api.v1.models.base_model import BaseTableModel
 
 
-class Newsletter(BaseTableModel):
-    """
-    Newsletter db model
-    """
-    __tablename__ = 'newsletters'
+# class Newsletter(BaseTableModel):
+#     """
+#     Newsletter db model
+#     """
+#     __tablename__ = 'newsletters'
 
-    email = Column(String(150), unique=True, nullable=False)
+#     email = Column(String(150), unique=True, nullable=False)
