+++ api/v1/models/profile.py
@@ -1,24 +1,24 @@
-#!/usr/bin/env python3
-""" The Profile model
-"""
-from sqlalchemy import (
-        Column,
-        String,
-        Text,
-        ForeignKey,
-        )
-from sqlalchemy.orm import relationship
-from api.v1.models.base import Base
-from api.v1.models.base_model import BaseTableModel
+# #!/usr/bin/env python3
+# """ The Profile model
+# """
+# from sqlalchemy import (
+#         Column,
+#         String,
+#         Text,
+#         ForeignKey,
+#         )
+# from sqlalchemy.orm import relationship
+# from api.v1.models.base import Base
+# from api.v1.models.base_model import BaseTableModel
 
 
 
-class Profile(BaseTableModel):
-    __tablename__ = 'profiles'
+# class Profile(BaseTableModel):
+#     __tablename__ = 'profiles'
 
-    user_id = Column(String, ForeignKey('users.id'), unique=True, nullable=False)
-    bio = Column(Text, nullable=True)
-    phone_number = Column(String(50), nullable=True)
-    avatar_url = Column(String(100), nullable=True)
+#     user_id = Column(String, ForeignKey('users.id'), unique=True, nullable=False)
+#     bio = Column(Text, nullable=True)
+#     phone_number = Column(String(50), nullable=True)
+#     avatar_url = Column(String(100), nullable=True)
 
-    user = relationship("User", back_populates="profile")
+#     user = relationship("User", back_populates="profile")
