+++ api/v1/models/user.py
@@ -18,7 +18,7 @@ from sqlalchemy import (
         )
 from sqlalchemy.orm import relationship
 from datetime import datetime
-from api.v1.models.base import Base, user_organization_association, user_role_association
+from api.v1.models.base import Base, user_organization_association
 from api.v1.models.base_model import BaseTableModel
 from sqlalchemy.dialects.postgresql import UUID
 
@@ -26,19 +26,17 @@ from sqlalchemy.dialects.postgresql import UUID
 class User(BaseTableModel, Base):
     __tablename__ = 'users'
 
-    username = Column(String(50), unique=True, nullable=False)
-    email = Column(String(100), unique=True, nullable=False)
-    password = Column(String(255), nullable=False)
-    first_name = Column(String(50))
-    last_name = Column(String(50))
-    is_active = Column(Boolean, server_default=text('true'))
-    is_admin = Column(Boolean, server_default=text('false'))
-    is_deleted = Column(Boolean, server_default=text('false'))
+    username = Column(String, unique=True, nullable=False)
+    email = Column(String, unique=True, nullable=False)
+    password = Column(String, nullable=True)
+    first_name = Column(String, nullable=True)
+    last_name = Column(String, nullable=True)
+    is_active = Column(Boolean, default=True)
+    is_super_admin = Column(Boolean, default=False)
 
-    profile = relationship("Profile", uselist=False, back_populates="user", cascade="all, delete-orphan")
+    # profile = relationship("Profile", uselist=False, back_populates="user", cascade="all, delete-orphan")
     organizations = relationship("Organization", secondary=user_organization_association, back_populates="users")
-    roles = relationship('Role', secondary=user_role_association, back_populates='users')
-
+    roles = relationship("OrgRole", back_populates="user", cascade="all, delete-orphan")
     def to_dict(self):
         obj_dict = super().to_dict()
         obj_dict.pop("password")
@@ -47,9 +45,3 @@ class User(BaseTableModel, Base):
 
     def __str__(self):
         return self.email
-
-class WaitlistUser(BaseTableModel):
-    __tablename__ = 'waitlist_users'
-
-    email = Column(String(100), unique=True, nullable=False)
-    full_name = Column(String(100), nullable=False)
